+++ next.config.ts
@@ -50,11 +50,9 @@ const withMDX = createMDX({
   },
 })
 
-export default process.env.POSTHOG_API_KEY && process.env.POSTHOG_ENV_ID
-  ? withPostHogConfig(withMDX(nextConfig), {
-      personalApiKey: process.env.POSTHOG_API_KEY,
-      envId: process.env.POSTHOG_ENV_ID,
-      host: process.env.NEXT_PUBLIC_POSTHOG_HOST,
-      sourcemaps: { enabled: process.env.VERCEL_ENV === 'production' },
-    })
-  : withMDX(nextConfig)
+export default withPostHogConfig(withMDX(nextConfig), {
+  personalApiKey: process.env.POSTHOG_API_KEY!,
+  envId: process.env.POSTHOG_ENV_ID!,
+  host: process.env.NEXT_PUBLIC_POSTHOG_HOST,
+  sourcemaps: { enabled: process.env.VERCEL_ENV === 'production' },
+})
