+++ public/compare/compare-aws.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:13dccaa22c3a5632e4877b97c1f29fc782e605bcae2970a41e87d7cafcb57e9a
+size 283322
