+++ public/compare/compare-gcp.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cbb300534782474e7eac2cdca12647ba71360327b07bf4dddd633f63296bc912
+size 267662
