+++ public/compare/compare-tinfoil.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e92c620b522fcd4990fe0f310f483eb1f96fa76a5ae9a1bcb5be2c665345351e
+size 263671
