+++ src/app/globals.css
@@ -1,7 +1,7 @@
-@import "tailwindcss";
-@import "tw-animate-css";
-@import "./theme/colors.css";
-@import "./theme/typography.css";
+@import 'tailwindcss';
+@import 'tw-animate-css';
+@import './theme/colors.css';
+@import './theme/typography.css';
 
 @custom-variant dark (&:is(.dark *));
 @plugin '@tailwindcss/typography';
