+++ src/app/sitemap.ts
@@ -1,15 +1,13 @@
 import type { MetadataRoute } from 'next'
 
 import { env } from '@/env'
-// Temporarily disabled - requires Notion API
-// import { getRecentPosts, retrieveTags } from '@/lib/post'
+import { getRecentPosts, retrieveTags } from '@/lib/post'
 
 const BASE_URL = `https://${env.VERCEL_PROJECT_PRODUCTION_URL}`
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
-  // Temporarily disabled - requires Notion API
-  const tags: string[] = [] // await retrieveTags()
-  const posts: any[] = [] // await getRecentPosts(100)
+  const tags = await retrieveTags()
+  const posts = await getRecentPosts(100)
 
   const staticPages: MetadataRoute.Sitemap = [
     {
