+++ src/components/DotBackground.tsx
@@ -1,20 +1,9 @@
 import { cn } from '@/lib/utils'
 
-export default function DotBackground({
-  dotColor,
-  bgColor,
-  className,
-}: {
-  dotColor: string
-  bgColor: string
-  className?: string
-}) {
+export default function DotBackground({ dotColor, bgColor, className }: { dotColor: string, bgColor: string, className?: string }) {
   return (
     <div
-      className={cn(
-        'absolute inset-0 -z-10 h-full w-full cursor-none pointer-events-none',
-        className,
-      )}
+      className={cn("absolute inset-0 -z-10 h-full w-full cursor-none pointer-events-none", className)}
       style={{
         backgroundColor: bgColor,
         backgroundImage: `radial-gradient(${dotColor} 2px, ${bgColor} 2px)`,
@@ -23,3 +12,4 @@ export default function DotBackground({
     />
   )
 }
+
