+++ src/components/nav.tsx
@@ -7,7 +7,6 @@ import {
   Brain,
   Coins,
   Cpu,
-  GitCompare,
   Gpu,
   GraduationCap,
   Lightbulb,
@@ -113,12 +112,6 @@ const menu = [
         description: 'Latest insights and updates from the Phala ecosystem',
         icon: <BookOpen className="size-5 shrink-0" />,
         url: '/blog',
-      },
-      {
-        title: 'Compare Alternatives',
-        description: 'See how Phala stacks up against other platforms',
-        icon: <GitCompare className="size-5 shrink-0" />,
-        url: '/compare',
       },
     ],
   },
