+++ src/components/roadmap/style.css
@@ -5,7 +5,7 @@
   display: block;
 }
 
-@media (min-aspect-ratio: 1 / 1) {
+@media (min-aspect-ratio: 1/1) {
   .roadmap-canvas {
     padding-top: 7rem;
   }
